+++ build_runner/lib/src/build_plan/build_phase_creator.dart
@@ -192,7 +192,7 @@ class BuildPhaseCreator {
             options: options,
             hideOutput: builderDefinition.hideOutput,
             isOptional: builderDefinition.isOptional,
-            writesParts: builderDefinition.writesParts,
+            addsToLibrary: builderDefinition.addsToLibrary,
           ),
         );
       }