+++ build_test/lib/src/test_builder.dart
@@ -189,7 +189,7 @@ Future<TestBuilderResult> testBuilders(
   Resolvers? resolvers,
   Set<Builder> optionalBuilders = const {},
   Set<Builder> visibleOutputBuilders = const {},
-  Set<Builder> writesPartsBuilders = const {},
+  Set<Builder> addsToLibraryBuilders = const {},
   Map<Builder, List<String>> appliesBuilders = const {},
   bool testingBuilderConfig = true,
   TestReaderWriter? readerWriter,
@@ -199,7 +199,7 @@ Future<TestBuilderResult> testBuilders(
   final builderFactories = <BuilderFactory>[];
   final optionalBuilderFactories = Set<BuilderFactory>.identity();
   final visibleOutputBuilderFactories = Set<BuilderFactory>.identity();
-  final writesPartsBuilderFactories = Set<BuilderFactory>.identity();
+  final addsToLibraryBuilderFactories = Set<BuilderFactory>.identity();
   final appliesBuildersToFactories = <BuilderFactory, List<String>>{};
   for (final builder in builders) {
     Builder builderFactory(_) => builder;
@@ -210,8 +210,8 @@ Future<TestBuilderResult> testBuilders(
     if (visibleOutputBuilders.contains(builder)) {
       visibleOutputBuilderFactories.add(builderFactory);
     }
-    if (writesPartsBuilders.contains(builder)) {
-      writesPartsBuilderFactories.add(builderFactory);
+    if (addsToLibraryBuilders.contains(builder)) {
+      addsToLibraryBuilderFactories.add(builderFactory);
     }
     if (appliesBuilders.containsKey(builder)) {
       appliesBuildersToFactories[builderFactory] = appliesBuilders[builder]!;
@@ -235,7 +235,7 @@ Future<TestBuilderResult> testBuilders(
     resolvers: resolvers,
     optionalBuilderFactories: optionalBuilderFactories,
     visibleOutputBuilderFactories: visibleOutputBuilderFactories,
-    writesPartsBuilderFactories: writesPartsBuilderFactories,
+    addsToLibraryBuilderFactories: addsToLibraryBuilderFactories,
     appliesBuilders: appliesBuildersToFactories,
     testingBuilderConfig: testingBuilderConfig,
     readerWriter: readerWriter,
@@ -326,7 +326,7 @@ Future<TestBuilderResult> testBuilderFactories(
   Resolvers? resolvers,
   Set<BuilderFactory> optionalBuilderFactories = const {},
   Set<BuilderFactory> visibleOutputBuilderFactories = const {},
-  Set<BuilderFactory> writesPartsBuilderFactories = const {},
+  Set<BuilderFactory> addsToLibraryBuilderFactories = const {},
   Map<BuilderFactory, List<String>> appliesBuilders = const {},
   bool testingBuilderConfig = true,
   TestReaderWriter? readerWriter,
@@ -436,7 +436,7 @@ Future<TestBuilderResult> testBuilderFactories(
           autoApply: build_config.AutoApply.allPackages,
           isOptional: optionalBuilderFactories.contains(builderFactory),
           hideOutput: !visibleOutputBuilderFactories.contains(builderFactory),
-          writesParts: writesPartsBuilderFactories.contains(builderFactory),
+          addsToLibrary: addsToLibraryBuilderFactories.contains(builderFactory),
           appliesBuilders: appliesBuilders[builderFactory] ?? const [],
         ),
         applyToPackages: inputPackages,
@@ -607,7 +607,7 @@ class _ApplyBuilderDefinitionToPackages implements BuilderDefinition {
   String get package => delegate.package;
 
   @override
-  bool get writesParts => delegate.writesParts;
+  bool get addsToLibrary => delegate.addsToLibrary;
 
   @override
   bool get hideOutput => delegate.hideOutput;