+++ build_config/CHANGELOG.md
@@ -1,6 +1,10 @@
+## 1.4.0-wip
+
+- Add boolean `adds_to_library` builder configuration that grants access to new
+   builder feature `BuildStep.librarySourceSink`.
+
 ## 1.3.1
 
-- Add boolean `writes_parts` builder configuration to opt in to writing parts via `BuildStep.partWriter`.
 - Document that `--define` values are parsed as JSON with a fallback.
 - Require Dart 3.8.0.